BackendFastApi/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
transformers==4.48.0
torch==2.5.1
numpy==2.2.2
setuptools==75.8.0
python-multipart==0.0.20
